refactor: 금액 관련 타입을 Int에서 Long으로 변경 및 CartRepository 메서드 추가

+++ app/src/main/java/woowacourse/shopping/repository/network/LegacyNetworkProductRepository.kt
@@ -46,7 +46,7 @@ class LegacyNetworkProductRepository(
                 Product(
                     id = jsonObject.getLong("id"),
                     name = jsonObject.getString("name"),
-                    price = Money(jsonObject.getInt("price")),
+                    price = Money(jsonObject.getLong("price")),
                     imageUrl = jsonObject.getString("imageUrl"),
                 ),
             )
